+++ CHANGELOG.md
@@ -9,6 +9,7 @@ New developments that use __graphql goal/task__ should use the parameters below,
 
 Both mode (client and server):
 * Upgrade of graphql-java from v14.0 to v15.0
+* Main improvement for this: the plugin now accepts that an interface implements one or more interfaces
 
 Server mode:
 * The generated code would not compile if a GraphQL interface is defined, but not used, in the given GraphQL schema